api_service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5
aiohttp>=3.8.0
PyYAML>=5.4.0
sqlalchemy>=1.4.0
alembic>=1.7.0
httpx>=0.24.0
email-validator>=2.0.0
opencv-python-headless>=4.8.0
pymysql>=1.0.2
cryptography>=36.0.0
mysqlclient>=2.1.0
paho-mqtt>=2.1.0
psutil>=5.9.0
GPUtil>=1.4.0
python-ffmpeg-video-streaming>=0.1.15
redis[hiredis]>=5.0.1
aioredis>=2.0.0
